Separate JS files for logic and UI into two files (timer.js and index.js).
Import the index.js file as a module into the HTML to avoid ES6-related error.
