+++ src/timer.html
@@ -7,7 +7,7 @@
     <title>Timer</title>
 </head>
 <body>
-    
+
     <svg class="timer" width="90vw" height="2.5rem">
         <rect class="timer_background" width="100%" height="40" />
         <rect class="timer_progress-bar" width="100%" height="40" />
@@ -15,7 +15,7 @@
         <text class="timer_remaining-time" x="1%" y="60%">0</text>
     </svg>
 
-    <script src="timer.js"></script>
-    
+    <script src="index.js" type="module"></script>
+
 </body>
 </html>
